Split resolver declaration tests
